requirements.txt
django
django-bootstrap-v5
django-jquery
django-bootstrap-datepicker-plus